Refecatored the MarkupList widget to its own class for easy re-use. WIP
